requirements.txt
pyluach~=2.2
pydantic~=1.10
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
pytest~=7.2.2